// src/app/pages/users/dto/users.dto.ts
export class UserListDto {
    id: number;
    name: string;
    lastName: string;
    phoneNumber: number;
    balance: number;
    address: string;
    createdDate: Date;
    countryname: string;
    active: boolean;
    email: string;
    date: Date;
    total: number;
    seller: boolean;
    seller_percentage: number;
    limite: any;
}